MLIR: Fix batched forward mode (#2578)
* MLIR: Fix batched forward mode

* fix

+++ enzyme/Enzyme/MLIR/Passes/RemovalUtils.h
@@ -492,7 +492,7 @@ public:
             }
 
             SmallVector<Value> dynSizes;
-            for (int i = inductionVariable.size(); i < dynamicDims.size();
+            for (size_t i = inductionVariable.size(); i < dynamicDims.size();
                  ++i) {
               dynSizes.push_back(dynamicDims[i]);
             }
@@ -727,7 +727,7 @@ public:
           sizes.append(shape.begin(), shape.end());
 
           SmallVector<Value> dynSizes;
-          for (int i = reversedIndex.size(); i < newShape.size(); ++i) {
+          for (size_t i = reversedIndex.size(); i < newShape.size(); ++i) {
             // we use memref.dim here to know the size, hopefully further
             // optimization/canonicalizations can just forward the right size
             // here.
